extract component where users selects airports into a separate component
