+++ screens/Homescreen.tsx
@@ -1,21 +1,20 @@
 import { useNavigation } from "@react-navigation/native";
 import { NativeStackNavigationProp } from "@react-navigation/native-stack";
 import React, { Fragment, useEffect, useState } from "react";
-import { FlatList, Keyboard, View } from "react-native";
+import { FlatList, View } from "react-native";
 import MapView, { Marker, Polygon } from "react-native-maps";
 import {
   Button,
   Icon,
   IconButton,
-  List,
   Snackbar,
   Surface,
   Text,
-  TextInput,
   useTheme,
 } from "react-native-paper";
 import { BottomModal } from "../components/BottomModal";
 import { MyFAB } from "../components/FAB";
+import TripModal from "../components/TripModal";
 import { LATITUDE, LONGITUDE } from "../constants/constants";
 import { Airport, airports } from "../data/airports";
 import { getFavorites, updateFavorites } from "../data/store";
@@ -41,10 +40,6 @@ export default function Homescreen() {
   const depAirport = useFlightStore((s) => s.depAirport);
   const arrAirport = useFlightStore((s) => s.arrAirport);
   const [selectedAirport, setSelectedAirport] = useState<Airport | null>(null);
-  const [activeField, setActiveField] = useState<
-    "departure" | "arrival" | null
-  >(null);
-  const [search, setSearch] = useState("");
   const [query, expanded, setExpanded] = useLazyQuery(["sigmets"], getSigmets);
   const [snackbar, setSnackbar] = useState(false);
   const [sigmetSnack, setSigmetSnack] = useState(false);
@@ -61,16 +56,6 @@ export default function Homescreen() {
 
   const navigation =
     useNavigation<NativeStackNavigationProp<RootStackParamList>>();
-
-  const filteredAirports = airports.filter((a) => {
-    if (!search) {
-      return favorites.includes(a.icao);
-    }
-    return (
-      a.name.toLowerCase().includes(search.toLowerCase()) ||
-      a.icao.toLowerCase().includes(search.toLowerCase())
-    );
-  });
 
   const handleToggleFavorite = async (icao: string) => {
     let updated: string[];
@@ -105,15 +90,6 @@ export default function Homescreen() {
 
   const handleNewTrip = () => {
     setShowModal(true);
-  };
-
-  const switchDepartureArrival = () => {
-    const dep = departure;
-    const depAir = depAirport;
-    setDeparture(arrival);
-    setDepAirport(arrAirport);
-    setArrival(dep);
-    setArrAirport(depAir);
   };
 
   const handleMarkerPressed = (airport: Airport) => {
@@ -178,167 +154,23 @@ export default function Homescreen() {
             ) : null
           )}
       </MapView>
-      <BottomModal visible={showModal} onClose={() => setShowModal(false)}>
-        <View
-          style={{
-            flexDirection: "row",
-            alignItems: "center",
-            justifyContent: "space-between",
-          }}
-        >
-          <Text variant="titleLarge">Prepare Trip</Text>
-          <IconButton
-            icon="close"
-            iconColor={colors.onBackground}
-            size={20}
-            onPress={() => setShowModal(false)}
-          />
-        </View>
-        <View style={{ position: "relative" }}>
-          <TextInput
-            label="Departure airport"
-            value={departure ?? depAirport?.name ?? ""}
-            onFocus={() => {
-              setActiveField("departure");
-              setSearch("");
-            }}
-            onChangeText={(text) => {
-              setDeparture(text);
-              setSearch(text);
-            }}
-            mode="outlined"
-            outlineColor={colors.onBackground}
-            activeOutlineColor={colors.primary}
-            textColor={colors.onBackground}
-            style={{ marginTop: 10 }}
-            left={
-              <TextInput.Icon
-                icon="airplane-takeoff"
-                size={20}
-                color={colors.tertiary}
-              />
-            }
-            right={
-              <TextInput.Icon
-                icon="close"
-                size={20}
-                color={colors.onBackground}
-                onPress={() => {
-                  setDeparture(null);
-                  setDepAirport(null);
-                }}
-              />
-            }
-          />
-          <TextInput
-            label="Arrival airport"
-            value={arrival ?? arrAirport?.name ?? ""}
-            disabled={depAirport === null}
-            onFocus={() => {
-              setActiveField("arrival");
-              setSearch("");
-            }}
-            onChangeText={(text) => {
-              setArrival(text);
-              setSearch(text);
-            }}
-            mode="outlined"
-            outlineColor={colors.onBackground}
-            activeOutlineColor={colors.primary}
-            textColor={colors.onBackground}
-            style={{ marginTop: 10 }}
-            left={
-              <TextInput.Icon
-                icon="airplane-landing"
-                size={20}
-                color={colors.tertiary}
-              />
-            }
-            right={
-              <TextInput.Icon
-                icon="close"
-                size={20}
-                color={colors.onBackground}
-                onPress={() => {
-                  setArrival(null);
-                  setArrAirport(null);
-                }}
-              />
-            }
-          />
-          <IconButton
-            icon="arrow-up-down"
-            iconColor={colors.primary}
-            size={30}
-            onPress={switchDepartureArrival}
-            style={{
-              position: "absolute",
-              top: 45,
-              right: 50,
-              zIndex: 2,
-              backgroundColor: colors.background,
-            }}
-          />
-        </View>
-        {activeField && (
-          <FlatList
-            style={{ flex: 1 }}
-            data={filteredAirports}
-            keyboardShouldPersistTaps="handled"
-            keyExtractor={(item: Airport) => item.icao}
-            renderItem={({ item }: { item: Airport }) => (
-              <List.Item
-                title={item.icao}
-                titleStyle={{ color: colors.primary }}
-                description={item.name}
-                descriptionStyle={{ color: colors.onBackground }}
-                left={(props) => (
-                  <List.Icon
-                    {...props}
-                    icon="airplane"
-                    color={colors.tertiary}
-                  />
-                )}
-                right={() => (
-                  <IconButton
-                    icon={
-                      favorites?.includes(item.icao) ? "star" : "star-outline"
-                    }
-                    onPress={() => handleToggleFavorite(item.icao)}
-                    size={20}
-                    iconColor={colors.tertiary}
-                  />
-                )}
-                onPress={() => {
-                  Keyboard.dismiss();
-                  if (activeField === "departure") {
-                    setDeparture(item.name);
-                    setDepAirport(item);
-                  } else {
-                    setArrival(item.name);
-                    setArrAirport(item);
-                    setActiveField(null);
-                  }
-                }}
-              />
-            )}
-            ListEmptyComponent={
-              <Text variant="titleMedium" style={{ marginTop: 8 }}>
-                Begin typing to see airports
-              </Text>
-            }
-          />
-        )}
-        {depAirport && (
-          <Button
-            icon="arrow-right-bold"
-            onPress={handleGoToBrief}
-            style={{ marginVertical: 50 }}
-          >
-            Go To Brief
-          </Button>
-        )}
-      </BottomModal>
+      <TripModal
+        visible={showModal}
+        onClose={() => setShowModal(false)}
+        airports={airports}
+        favorites={favorites}
+        colors={colors}
+        depAirport={depAirport}
+        arrAirport={arrAirport}
+        departure={departure}
+        arrival={arrival}
+        setDeparture={setDeparture}
+        setArrival={setArrival}
+        setDepAirport={setDepAirport}
+        setArrAirport={setArrAirport}
+        handleToggleFavorite={handleToggleFavorite}
+        handleGoToBrief={handleGoToBrief}
+      />
       <MyFAB
         visible={true}
         extended={true}
